+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 * [#987](https://github.com/bbatsov/projectile/issues/987): projectile-ag ignores ag-ignore-list when projectile-project-vcs is git
 * [#1119](https://github.com/bbatsov/projectile/issues/1119): File search ignores non-root dirs if prefixed with "*"
 * Treat members of `projectile-globally-ignored-file-suffixes` as file name suffixes (previous treat as file extensions).
+* Ensure project roots are added as directory names to avoid near-duplicate projects, e.g. "~/project/" and "~/project".
 
 ### Bugs fixed